+++ src/main/java/com/dabomstew/pkrandom/SysConstants.java
@@ -32,7 +32,7 @@ public class SysConstants {
     public static final String AUTOUPDATE_URL =
             "http://pokehacks.dabomstew.com/randomizer/autoupdate/";
     public static final String WEBSITE_URL = "http://pokehacks.dabomstew.com/randomizer/";
-    public static final int UPDATE_VERSION = 2010;
+    public static final int UPDATE_VERSION = 2020;
     public static final String ROOT_PATH = getRootPath();
     public static final String LINE_SEP = System.getProperty("line.separator");
     public static final String customNamesFile = "customnames.rncn";