+++ src/main/java/com/dabomstew/pkrandom/settings/Settings.java
@@ -56,7 +56,7 @@ import com.dabomstew.pkrandom.romhandlers.RomHandler;
 
 public class Settings {
 
-    public static final int VERSION = 201;
+    public static final int VERSION = 202;
 
     public static final int LENGTH_OF_SETTINGS_DATA = 46;